Add files via upload
This module deals with analysing data and predicting the Customer Lifetime Value  in $ of a particular customer by taking into account purchase and attribute history of thousands of existing records. We use linear regression because we have concrete output data to train our model against. In this particular case we managed to achieve an accuracy of 91.59%